requirements.txt
fastapi
uvicorn[standard]
gunicorn
pandas
selenium
websocket-client
sqlalchemy
APScheduler
unicorn_binance_websocket_api
python-dotenv
pydantic
requests
fastapi-mail
fastapi_utils
pandas_ta
asyncio
tqdm
ccxt
python-binance
binance-connector
aiohttp